microservices/text-analyzer/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-multipart
requests
psycopg2-binary
transformers
